+++ R/Balance.R
@@ -61,6 +61,9 @@ computeMeansPerGroup <- function(cohorts, cohortMethodData, covariateFilter) {
   useWeighting <- (hasStrata && any(stratumSize %>% pull(.data$n) > 1)) ||
     (!hasStrata && hasIptw)
 
+  # By definition:
+  sumW <- 1
+
   covariates <- filterCovariates(cohortMethodData$covariates, cohortMethodData$covariateRef, covariateFilter)
 
   if (useWeighting) {
@@ -70,73 +73,120 @@ computeMeansPerGroup <- function(cohorts, cohortMethodData, covariateFilter) {
         mutate(weight = 1 / .data$n) %>%
         inner_join(cohorts, by = c("stratumId", "treatment")) %>%
         select("rowId", "treatment", "weight")
+      # Overall weight is for computing mean and SD across T and C
+      overallW <-  stratumSize %>%
+        group_by(.data$stratumId) %>%
+        summarise(weight = 1 / sum(.data$n, na.rm = TRUE)) %>%
+        ungroup() %>%
+        inner_join(cohorts, by = c("stratumId")) %>%
+        select("rowId", "weight")
     } else {
       w <- cohorts %>%
         mutate(weight = .data$iptw) %>%
         select("rowId", "treatment", "weight")
+      overallW <- w
     }
     # Normalize so sum(weight) == 1 per treatment arm:
     wSum <- w %>%
       group_by(.data$treatment) %>%
       summarize(wSum = sum(.data$weight, na.rm = TRUE)) %>%
       ungroup()
+    overallWSum <- overallW %>%
+      summarize(overallWSum = sum(.data$weight, na.rm = TRUE)) %>%
+      pull()
 
     cohortMethodData$w <- w %>%
       inner_join(wSum, by = "treatment") %>%
       mutate(weight = .data$weight / .data$wSum) %>%
       select("rowId", "treatment", "weight")
 
-    # By definition:
-    sumW <- 1
+    cohortMethodData$overallW <- overallW %>%
+      mutate(overallWeight = .data$weight / overallWSum) %>%
+      select("rowId", "overallWeight")
 
     # Note: using abs() because due to rounding to machine precision number can become slightly negative:
     result <- covariates %>%
       inner_join(cohortMethodData$w, by = c("rowId")) %>%
+      inner_join(cohortMethodData$overallW, by = c("rowId")) %>%
       group_by(.data$covariateId, .data$treatment) %>%
       summarise(
         sum = sum(as.numeric(.data$covariateValue), na.rm = TRUE),
         mean = sum(.data$weight * as.numeric(.data$covariateValue), na.rm = TRUE),
+        overallMean = sum(.data$overallWeight * as.numeric(.data$covariateValue), na.rm = TRUE),
         sumSqr = sum(.data$weight * as.numeric(.data$covariateValue)^2, na.rm = TRUE),
-        sumWSqr = sum(.data$weight^2, na.rm = TRUE)
+        sumWSqr = sum(.data$weight^2, na.rm = TRUE),
+        overallSumSqr = sum(.data$overallWeight * as.numeric(.data$covariateValue)^2, na.rm = TRUE),
+        overallSumWSqr = sum(.data$overallWeight^2, na.rm = TRUE),
+        .groups = "drop"
       ) %>%
       mutate(sd = sqrt(abs(.data$sumSqr - .data$mean^2) * sumW / (sumW^2 - .data$sumWSqr))) %>%
       ungroup() %>%
-      select("covariateId", "treatment", "sum", "mean", "sd") %>%
+      select("covariateId", "treatment", "sum", "mean", "sd", "overallMean", "overallSumSqr", "overallSumWSqr") %>%
       collect()
 
     cohortMethodData$w <- NULL
+    cohortMethodData$overallW <- NULL
   } else {
+    # Don't use weighting
     cohortCounts <- cohorts %>%
       group_by(.data$treatment) %>%
       count()
+    overallCount <- cohorts %>%
+      count() %>%
+      pull()
 
     result <- covariates %>%
       inner_join(select(cohorts, "rowId", "treatment"), by = "rowId") %>%
       group_by(.data$covariateId, .data$treatment) %>%
       summarise(
         sum = sum(as.numeric(.data$covariateValue), na.rm = TRUE),
-        sumSqr = sum(as.numeric(.data$covariateValue)^2, na.rm = TRUE)
+        sumSqr = sum(as.numeric(.data$covariateValue)^2, na.rm = TRUE),
+        overallSumWSqr = sum(1 / overallCount^2, na.rm = TRUE),
+        .groups = "drop"
       ) %>%
       inner_join(cohortCounts, by = "treatment") %>%
       mutate(
         sd = sqrt((.data$sumSqr - (.data$sum^2 / .data$n)) / .data$n),
-        mean = .data$sum / .data$n
+        mean = .data$sum / .data$n,
+        overallMean = .data$sum / overallCount,
+        overallSumSqr = .data$sumSqr / overallCount
       ) %>%
       ungroup() %>%
-      select("covariateId", "treatment", "sum", "mean", "sd") %>%
+      select("covariateId", "treatment", "sum", "mean", "sd", "overallMean", "overallSumSqr", "overallSumWSqr") %>%
       collect()
   }
   target <- result %>%
     filter(.data$treatment == 1) %>%
-    select("covariateId", sumTarget = "sum", meanTarget = "mean", sdTarget = "sd")
+    select("covariateId",
+           sumTarget = "sum",
+           meanTarget = "mean",
+           sdTarget = "sd",
+           overallMeanTarget = "overallMean",
+           overallSumSqrTarget = "overallSumSqr",
+           overallSumWSqrTarget = "overallSumWSqr")
 
   comparator <- result %>%
     filter(.data$treatment == 0) %>%
-    select("covariateId", sumComparator = "sum", meanComparator = "mean", sdComparator = "sd")
+    select("covariateId",
+           sumComparator = "sum",
+           meanComparator = "mean",
+           sdComparator = "sd",
+           overallMeanComparator = "overallMean",
+           overallSumSqrComparator = "overallSumSqr",
+           overallSumWSqrComparator = "overallSumWSqr")
 
   result <- target %>%
     full_join(comparator, by = "covariateId") %>%
-    mutate(sd = sqrt((.data$sdTarget^2 + .data$sdComparator^2) / 2))
+    mutate(mean = .data$overallMeanTarget + .data$overallMeanComparator,
+           overallSumSqr = .data$overallSumSqrTarget + .data$overallSumSqrComparator,
+           overallSumWSqr = .data$overallSumWSqrTarget + .data$overallSumWSqrComparator) %>%
+    mutate(sd = sqrt(abs(.data$overallSumSqr - .data$mean^2) * sumW / (sumW^2 - .data$overallSumWSqr))) %>%
+    select(-"overallMeanTarget",
+           -"overallMeanComparator",
+           -"overallSumSqrTarget",
+           -"overallSumSqrComparator",
+           -"overallSumWSqrTarget",
+           -"overallSumWSqrComparator")
 
   return(result)
 }
@@ -171,7 +221,8 @@ computeMeansPerGroup <- function(cohorts, cohortMethodData, covariateFilter) {
 #'
 #' @return
 #' Returns a tibble describing the covariate balance before and after PS adjustment,
-#' with one row per covariate and the following columns:
+#' with one row per covariate, with the same data as the `covariateRef` table in the `CohortMethodData` object,
+#' and the following additional columns:
 #'
 #' - beforeMatchingMeanTarget: The (weighted) mean value in the target before PS adjustment.
 #' - beforeMatchingMeanComparator: The (weighted) mean value in the comparator before PS adjustment.
@@ -179,6 +230,7 @@ computeMeansPerGroup <- function(cohorts, cohortMethodData, covariateFilter) {
 #' - beforeMatchingSumComparator: The (weighted) sum value in the comparator before PS adjustment.
 #' - beforeMatchingSdTarget: The standard deviation of the value in the target before PS adjustment.
 #' - beforeMatchingSdComparator: The standard deviation of the value in the comparator before PS adjustment.
+#' - beforeMatchingMean: The mean of the value across target and comparator before PS adjustment.
 #' - beforeMatchingSd: The standard deviation of the value across target and comparator before PS adjustment.
 #' - afterMatchingMeanTarget: The (weighted) mean value in the target after PS adjustment.
 #' - afterMatchingMeanComparator: The (weighted) mean value in the comparator after PS adjustment.
@@ -186,6 +238,7 @@ computeMeansPerGroup <- function(cohorts, cohortMethodData, covariateFilter) {
 #' - afterMatchingSumComparator: The (weighted) sum value in the comparator after PS adjustment.
 #' - afterMatchingSdTarget: The standard deviation of the value in the target after PS adjustment.
 #' - afterMatchingSdComparator: The standard deviation of the value in the comparator after PS adjustment.
+#' - afterMatchingMean: The mean of the value across target and comparator after PS adjustment.
 #' - afterMatchingSd: The standard deviation of the value across target and comparator after PS adjustment.
 #' - beforeMatchingStdDiff: The standardized difference of means when comparing the target to
 #'                          the comparator before PS adjustment.
@@ -195,14 +248,17 @@ computeMeansPerGroup <- function(cohorts, cohortMethodData, covariateFilter) {
 #'                  before PS adjustment to the target after PS adjustment.
 #' - comparatorStdDiff: The standardized difference of means when comparing the comparator
 #'                      before PS adjustment to the comparator after PS adjustment.
+#'  -targetComparatorStdDiff:  The standardized difference of means when comparing the entire
+#'                             population before PS adjustment to the entire population after
+#'                             PS adjustment.
 #'
 #' The 'beforeMatchingStdDiff' and 'afterMatchingStdDiff' columns inform on the balance:
 #' are the target and comparator sufficiently similar in terms of baseline covariates to
 #' allow for valid causal estimation?
 #'
-#' The 'targetStdDiff' and 'comparatorStdDiff' columns inform on the generalizability:
-#' are the cohorts after PS adjustment sufficiently similar to the cohorts before adjustment
-#' to allow generalizing the findings to the original cohorts?
+#' The 'targetStdDiff', 'comparatorStdDiff', and 'targetComparatorStdDiff' columns inform on
+#' the generalizability: are the cohorts after PS adjustment sufficiently similar to the cohorts
+#' before adjustment to allow generalizing the findings to the original cohorts?
 #'
 #' @references
 #' Austin, P.C. (2008) Assessing balance in measured baseline covariates when using many-to-one
@@ -292,6 +348,7 @@ computeCovariateBalance <- function(population,
            beforeMatchingSumComparator = "sumComparator",
            beforeMatchingSdTarget = "sdTarget",
            beforeMatchingSdComparator = "sdComparator",
+           beforeMatchingMean = "mean",
            beforeMatchingSd = "sd")
   afterMatching <- afterMatching %>%
     select("covariateId",
@@ -301,7 +358,9 @@ computeCovariateBalance <- function(population,
            afterMatchingSumComparator = "sumComparator",
            afterMatchingSdTarget = "sdTarget",
            afterMatchingSdComparator = "sdComparator",
-           afterMatchingSd = "sd")
+           afterMatchingMean = "mean",
+           afterMatchingSd = "sd",
+           matches("overallMean"))
   balance <- beforeMatching %>%
     full_join(afterMatching, by = "covariateId") %>%
     inner_join(collect(cohortMethodData$covariateRef), by = "covariateId") %>%
@@ -329,12 +388,20 @@ computeCovariateBalance <- function(population,
         .data$beforeMatchingSdComparator == 0,
         0,
         (.data$beforeMatchingMeanComparator - .data$afterMatchingMeanComparator) / .data$beforeMatchingSdComparator
+      ),
+      targetComparatorStdDiff = if_else(
+        .data$beforeMatchingSd == 0,
+        0,
+        (.data$beforeMatchingMean - .data$afterMatchingMean) / .data$beforeMatchingSd
       )
 
     ) %>%
     arrange(desc(abs(.data$beforeMatchingStdDiff)))
-  # TODO: Compute generalizability across T and C
 
+  metaData <- attr(population, "metaData")
+  if (!is.null(metaData) && !is.null(metaData$targetEstimator)) {
+    attr(balance, "targetEstimator") <- metaData$targetEstimator
+  }
   delta <- Sys.time() - start
   message(paste("Computing covariate balance took", signif(delta, 3), attr(delta, "units")))
   return(balance)
@@ -698,4 +765,103 @@ plotCovariatePrevalence <- function(balance,
     ggplot2::ggsave(filename = fileName, plot = plot, width = 8, height = 4, dpi = 400)
   }
   return(plot)
+}
+
+#' Get information on generalizability
+#'
+#' @description
+#' to assess generalizability we compare the distribution of covariates before and after
+#' any (propensity score) adjustments. We compute the standardized difference of mean as
+#' our metric of generalizability. (Lipton et al., 2017)
+#'
+#' Depending on our target estimand, we need to consider a different base population for
+#' generalizability. For example, if we aim to estimate the average treatment effect in
+#' thetreated (ATT), our base population should be the target population, meaning we
+#' should consider the covariate distribution before and after PS adjustment in the target
+#' population only. By default this function will attempt to select the right base
+#' population based on what operations have been performed on the population. For example,
+#' if PS matching has been performed we assume the target estimand is the ATT, and the
+#' target population is selected as base.
+#'
+#' Requires running [computeCovariateBalance()]` first.
+#'
+#' @param balance       A data frame created by the `computeCovariateBalance` function.
+#' @param baseSelection The selection of the population to consider for generalizability.
+#'                      Options are "auto", "target", "comparator", and "both". The "auto"
+#'                      option will attempt to use the balance meta-data to pick the most
+#'                      appropriate population based on the target estimator.
+#'
+#' @return
+#' A tibble with the following columns:
+#'
+#' - covariateId: The ID of the covariate. Can be linked to the `covariates` and `covariateRef`
+#'   tables in the `CohortMethodData` object.
+#' - covariateName: The name of the covariate.
+#' - beforeMatchingMean: The mean covariate value before any (propensity score) adjustment.
+#' - afterMatchingMean: The mean covariate value after any (propensity score) adjustment.
+#' - stdDiff: The standardized difference of means between before and after adjustment.
+#'
+#' The tibble also has a 'baseSelection' attribute, documenting the base population used
+#' to assess generalizability.
+#'
+#' @references Tipton E, Hallberg K, Hedges LV, Chan W (2017) Implications of Small Samples
+#' for Generalization: Adjustments and Rules of Thumb, Eval Rev. Oct;41(5):472-505.
+#'
+#' @export
+getGeneralizabilityTable <- function(balance, baseSelection = "auto") {
+  errorMessages <- checkmate::makeAssertCollection()
+  checkmate::assertDataFrame(balance, add = errorMessages)
+  checkmate::assertCharacter(baseSelection, len = 1, add = errorMessages)
+  checkmate::assertChoice(baseSelection, c("auto", "target", "comparator",  "both"), add = errorMessages)
+  checkmate::reportAssertions(collection = errorMessages)
+
+  if (baseSelection == "auto") {
+    targetEstimator <- attr(balance, "targetEstimator")
+    if (is.null(targetEstimator)) {
+      stop("The baseSelection is set to 'auto' but the balance object does not contain a target estimator attribute. ",
+           "Please set the baseSelection manually.")
+    }
+    if (targetEstimator == "ate" | targetEstimator == "ato") {
+      baseSelection <- "both"
+      message("Selecting both target and comparator as base for generalizability")
+    } else if (targetEstimator == "att") {
+      baseSelection <- "target"
+      message("Selecting target as base for generalizability")
+    } else if (targetEstimator == "atu") {
+      baseSelection <- "comparator"
+      message("Selecting comparator as base for generalizability")
+    } else {
+      stop("Unkown target estimator: ", targetEstimator)
+    }
+  }
+  if (baseSelection == "target") {
+    generalizability <- balance %>%
+      mutate(absGeneralizabilityStdDiff = abs(.data$targetStdDiff)) %>%
+      arrange(desc(.data$absGeneralizabilityStdDiff)) %>%
+      select("covariateId",
+             "covariateName",
+             beforeMatchingMean = "beforeMatchingMeanTarget",
+             afterMatchingMean = "afterMatchingMeanTarget",
+             stdDiff = "targetStdDiff")
+  } else if (baseSelection == "comparator") {
+    generalizability <- balance %>%
+      mutate(absGeneralizabilityStdDiff = abs(.data$comparatorStdDiff)) %>%
+      arrange(desc(.data$absGeneralizabilityStdDiff)) %>%
+      select("covariateId",
+             "covariateName",
+             beforeMatchingMean = "beforeMatchingMeanComparator",
+             afterMatchingMean = "afterMatchingMeanComparator",
+             stdDiff = "comparatorStdDiff")
+  } else {
+    generalizability <- balance %>%
+      mutate(absGeneralizabilityStdDiff = abs(.data$targetComparatorStdDiff)) %>%
+      arrange(desc(.data$absGeneralizabilityStdDiff)) %>%
+      select("covariateId",
+             "covariateName",
+             "beforeMatchingMean",
+             "afterMatchingMean",
+             stdDiff = "targetComparatorStdDiff")
+  }
+  attr(generalizability, "baseSelection") <- baseSelection
+  return(generalizability)
 }