+++ R/RunAnalyses.R
@@ -474,7 +474,7 @@ runCmAnalyses <- function(connectionDetails,
       tasks <- split(subset, subset$sharedPsFile)
       cluster <- ParallelLogger::makeCluster(min(length(tasks), multiThreadingSettings$trimMatchStratifyThreads))
       ParallelLogger::clusterRequire(cluster, "CohortMethod")
-      dummy <- ParallelLogger::clusterApply(cluster, tasks, addPsToStudyPop, outputFolder = outputFolder)
+      dummy <- ParallelLogger::clusterApply(cluster, tasks, addPsToStudyPopForSubset, outputFolder = outputFolder)
       ParallelLogger::stopCluster(cluster)
     }
   }
@@ -691,7 +691,9 @@ runCmAnalyses <- function(connectionDetails,
       params <- list(
         cohortMethodDataFile = file.path(outputFolder, refRow$cohortMethodDataFile),
         prefilteredCovariatesFile = prefilteredCovariatesFile,
+        psFile =  file.path(outputFolder, refRow$psFile),
         sharedPsFile = file.path(outputFolder, refRow$sharedPsFile),
+        refitPsForEveryOutcome = refitPsForEveryOutcome,
         args = analysisRow,
         outcomeModelFile = file.path(outputFolder, refRow$outcomeModelFile)
       )
@@ -806,23 +808,35 @@ doFitSharedPsModel <- function(params, refitPsForEveryStudyPopulation) {
   return(NULL)
 }
 
-addPsToStudyPop <- function(subset, outputFolder) {
+addPsToStudyPopForSubset <- function(subset, outputFolder) {
   ps <- readRDS(file.path(outputFolder, subset$sharedPsFile[1]))
 
   addToStudyPop <- function(i) {
     refRow <- subset[i, ]
     studyPop <- readRDS(file.path(outputFolder, refRow$studyPopFile))
-    newMetaData <- attr(studyPop, "metaData")
-    newMetaData$psModelCoef <- attr(ps, "metaData")$psModelCoef
-    newMetaData$psModelPriorVariance <- attr(ps, "metaData")$psModelPriorVariance
-    idx <- match(studyPop$rowId, ps$rowId)
-    studyPop$propensityScore <- ps$propensityScore[idx]
-    studyPop$iptw <- ps$iptw[idx]
-    attr(studyPop, "metaData") <- newMetaData
+    studyPop <- addPsToStudyPopulation(studyPop, ps)
     saveRDS(studyPop, file.path(outputFolder, refRow$psFile))
     return(NULL)
   }
   plyr::l_ply(1:nrow(subset), addToStudyPop)
+}
+
+addPsToStudyPopulation <- function(studyPopulation, ps) {
+  # Merge meta-data
+  newMetaData <- attr(studyPopulation, "metaData")
+  psMetaData <-  attr(ps, "metaData")
+  missingColumns <- setdiff(names(psMetaData), names(newMetaData))
+  newMetaData <- append(newMetaData, psMetaData[missingColumns])
+  attr(studyPopulation, "metaData") <- newMetaData
+
+  # Merge data
+  missingColumns <- setdiff(names(ps), names(studyPopulation))
+  idx <- match(studyPopulation$rowId, ps$rowId)
+  studyPopulation <- bind_cols(
+    studyPopulation,
+    ps[idx, missingColumns]
+  )
+  return(studyPopulation)
 }
 
 applyTrimMatchStratify <- function(ps, arguments) {
@@ -944,12 +958,12 @@ doFitOutcomeModelPlus <- function(params) {
   studyPop <- do.call("createStudyPopulation", args)
 
   if (!is.null(params$args$createPsArgs)) {
-    # Add PS
-    ps <- getPs(params$sharedPsFile)
-    idx <- match(studyPop$rowId, ps$rowId)
-    studyPop$propensityScore <- ps$propensityScore[idx]
-    studyPop$iptw <- ps$iptw[idx]
-    ps <- studyPop
+    if (params$refitPsForEveryOutcome) {
+      ps <- getPs(params$psFile)
+    } else {
+      ps <- getPs(params$sharedPsFile)
+      ps <- addPsToStudyPopulation(studyPop, ps)
+    }
   } else {
     ps <- studyPop
   }
@@ -1287,41 +1301,45 @@ createReferenceTable <- function(cmAnalysisList,
 
   # Add shared covariate balance files (per target-comparator-analysis)
   args <- "computeSharedCovariateBalanceArgs"
-  normBalanceArgs <- function(balanceArgs) {
-    return(balanceArgs[args][!is.na(names(balanceArgs[args]))])
-  }
-  balanceArgsList <- unique(ParallelLogger::selectFromList(cmAnalysisList, args))
-  balanceArgsList <- balanceArgsList[sapply(
-    balanceArgsList,
-    function(balanceArgs) {
-      return(!is.null(balanceArgs$computeSharedCovariateBalanceArgs))
-    }
-  )]
-  balanceArgsList <- lapply(balanceArgsList, normBalanceArgs)
-  if (length(balanceArgsList) == 0) {
-    referenceTable$sharedBalanceId <- 0
+  if (refitPsForEveryOutcome) {
+    referenceTable$sharedBalanceFile <- ""
   } else {
-    sharedBalanceId <- sapply(cmAnalysisList, function(cmAnalysis) {
-      i <- which.list(balanceArgsList, normBalanceArgs(cmAnalysis))
-      if (is.null(i)) {
-        i <- 0
+    normBalanceArgs <- function(balanceArgs) {
+      return(balanceArgs[args][!is.na(names(balanceArgs[args]))])
+    }
+    balanceArgsList <- unique(ParallelLogger::selectFromList(cmAnalysisList, args))
+    balanceArgsList <- balanceArgsList[sapply(
+      balanceArgsList,
+      function(balanceArgs) {
+        return(!is.null(balanceArgs$computeSharedCovariateBalanceArgs))
       }
-      return(i)
-    })
-    analysisIdToBalanceArgsId <- tibble(analysisId = analyses$analysisId, sharedBalanceId = sharedBalanceId)
-    referenceTable <- inner_join(referenceTable, analysisIdToBalanceArgsId, by = "analysisId")
+    )]
+    balanceArgsList <- lapply(balanceArgsList, normBalanceArgs)
+    if (length(balanceArgsList) == 0) {
+      referenceTable$sharedBalanceId <- 0
+    } else {
+      sharedBalanceId <- sapply(cmAnalysisList, function(cmAnalysis) {
+        i <- which.list(balanceArgsList, normBalanceArgs(cmAnalysis))
+        if (is.null(i)) {
+          i <- 0
+        }
+        return(i)
+      })
+      analysisIdToBalanceArgsId <- tibble(analysisId = analyses$analysisId, sharedBalanceId = sharedBalanceId)
+      referenceTable <- inner_join(referenceTable, analysisIdToBalanceArgsId, by = "analysisId")
+    }
+    idx <- referenceTable$sharedBalanceId != 0
+    referenceTable$sharedBalanceFile <- ""
+    referenceTable$sharedBalanceFile[idx] <- .createsharedBalanceFileName(
+      loadId = referenceTable$loadArgsId[idx],
+      studyPopId = referenceTable$studyPopArgsId[idx],
+      psId = referenceTable$psArgsId[idx],
+      strataId = referenceTable$strataArgsId[idx],
+      sharedBalanceId = referenceTable$sharedBalanceId[idx],
+      targetId = referenceTable$targetId[idx],
+      comparatorId = referenceTable$comparatorId[idx]
+    )
   }
-  idx <- referenceTable$sharedBalanceId != 0
-  referenceTable$sharedBalanceFile <- ""
-  referenceTable$sharedBalanceFile[idx] <- .createsharedBalanceFileName(
-    loadId = referenceTable$loadArgsId[idx],
-    studyPopId = referenceTable$studyPopArgsId[idx],
-    psId = referenceTable$psArgsId[idx],
-    strataId = referenceTable$strataArgsId[idx],
-    sharedBalanceId = referenceTable$sharedBalanceId[idx],
-    targetId = referenceTable$targetId[idx],
-    comparatorId = referenceTable$comparatorId[idx]
-  )
 
   # Add covariate balance files (per target-comparator-analysis-outcome)
   args <- "computeCovariateBalanceArgs"
@@ -1373,7 +1391,7 @@ createReferenceTable <- function(cmAnalysisList,
     studyPopId = referenceTable$studyPopArgsId[idx],
     psId = referenceTable$psArgsId[idx],
     strataId = referenceTable$strataArgsId[idx],
-    sharedBalanceId = referenceTable$sharedBalanceId[idx],
+    balanceId = referenceTable$balanceId[idx],
     targetId = referenceTable$targetId[idx],
     comparatorId = referenceTable$comparatorId[idx],
     outcomeId = referenceTable$outcomeId[idx]
@@ -1495,11 +1513,16 @@ createReferenceTable <- function(cmAnalysisList,
 
   # Remove non-essential files for outcomes not of interest:
   idx <- !referenceTable$outcomeOfInterest
-  referenceTable$studyPopFile[idx] <- ""
-  referenceTable$psFile[idx] <- ""
   referenceTable$strataFile[idx] <- ""
   referenceTable$filteredForbalanceFile[idx] <- ""
   referenceTable$balanceFile[idx] <- ""
+  if (!refitPsForEveryOutcome) {
+    # If we're computing a PS per outcome it probably is a good idea to save it.
+    # For that it is more convenient to save the study population as well (so we can
+    # use the regular PS-model fitting routine)
+    referenceTable$studyPopFile[idx] <- ""
+    referenceTable$psFile[idx] <- ""
+  }
 
   # Remove rows that the user specified to exclude:
   if (!is.null(analysesToExclude)) {
@@ -1602,11 +1625,11 @@ createReferenceTable <- function(cmAnalysisList,
                                    studyPopId,
                                    psId,
                                    strataId,
-                                   sharedBalanceId,
+                                   balanceId,
                                    targetId,
                                    comparatorId,
                                    outcomeId) {
-  name <- sprintf("Balance_l%s_s%s_p%s_t%s_c%s_s%s_b%s_o%s.rds", loadId, studyPopId, psId, .f(targetId), .f(comparatorId), strataId, sharedBalanceId, .f(outcomeId))
+  name <- sprintf("Balance_l%s_s%s_p%s_t%s_c%s_s%s_b%s_o%s.rds", loadId, studyPopId, psId, .f(targetId), .f(comparatorId), strataId, balanceId, .f(outcomeId))
   return(name)
 }
 
@@ -1725,7 +1748,7 @@ summarizeResults <- function(referenceTable, outputFolder, mainFileName, interac
         seLogRr = if (is.null(coefficient)) NA else outcomeModel$outcomeModelTreatmentEstimate$seLogRr,
         llr = if (is.null(coefficient)) NA else outcomeModel$outcomeModelTreatmentEstimate$llr,
         mdrr = !!mdrr,
-        attritionFraction = !!attritionFraction
+        targetEstimator = outcomeModel$targetEstimator
       )
 
     mainResults[[i]] <- mainResult
@@ -1742,7 +1765,8 @@ summarizeResults <- function(referenceTable, outputFolder, mainFileName, interac
             ci95Ub = exp(outcomeModel$outcomeModelInteractionEstimates$logUb95[j]),
             p = !!p,
             logRr = outcomeModel$outcomeModelInteractionEstimates$logRr[j],
-            seLogRr = outcomeModel$outcomeModelInteractionEstimates$seLogRr[j]
+            seLogRr = outcomeModel$outcomeModelInteractionEstimates$seLogRr[j],
+            targetEstimator = outcomeModel$targetEstimator
           )
       }
     }